renamed modules and updated tests accordingly
