+++ src/codeoceanauxtools/get_read_direction.py
@@ -1,14 +1,15 @@
 #!/usr/bin/env python
 import sys
 
-from get_logger import LOGGER
-import co_aux_utils
+# from get_logger import LOGGER
+from codeoceanauxtools.get_logger import LOGGER
+import co_fastq
 
 
 def main(argv=None):
     if len(argv) == 1:
         sys.exit("You failed to provide a file name")
-    return co_aux_utils.get_read_direction(argv[1])
+    return co_fastq.get_read_direction(argv[1])
 
 
 if __name__ == "__main__":
