+++ src/codeoceanauxtools/get_read_pattern.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python
 import sys
 
-from get_logger import LOGGER
-import co_aux_utils
+# from get_logger import LOGGER
+from codeoceanauxtools.get_logger import LOGGER
+import co_fastq
 
 
 def main(argv=None):
@@ -10,8 +11,8 @@ def main(argv=None):
         LOGGER.error("You failed to provide a filename and direction")
         sys.exit("You failed to provide a filename and direction")
     elif len(argv) == 2:
-        return co_aux_utils.get_read_pattern(argv[1])
-    return co_aux_utils.get_read_pattern(argv[1], argv[2])
+        return co_fastq.get_read_pattern(argv[1])
+    return co_fastq.get_read_pattern(argv[1], argv[2])
 
 
 if __name__ == "__main__":
