+++ src/codeoceanauxtools/get_read_prefix.py
@@ -3,8 +3,9 @@ import re
 import sys
 from pathlib import Path
 
-from get_logger import LOGGER
-import co_aux_utils
+# from get_logger import LOGGER
+from codeoceanauxtools.get_logger import LOGGER
+import co_fastq
 
 
 def main(argv=None):
@@ -12,8 +13,8 @@ def main(argv=None):
         LOGGER.error("You failed to provide a filename")
         sys.exit("You failed to provide a filename")
     elif len(argv) == 2:
-        return co_aux_utils.get_prefix(argv[1])
-    return co_aux_utils.get_prefix(argv[1], argv[2])
+        return co_fastq.get_prefix(argv[1])
+    return co_fastq.get_prefix(argv[1], argv[2])
 
 
 if __name__ == "__main__":
