+++ tests/test_get_pipeline_confirm.py
@@ -1,6 +1,6 @@
 import os
 
-from src.codeoceanauxtools.utils.util import is_pipeline
+from src.codeoceanauxtools.co_utils import is_pipeline
 
 
 # Tests for is_pipeline
