+++ tests/test_get_read_prefix.py
@@ -1,4 +1,4 @@
-from src.codeoceanauxtools.utils.util import get_prefix
+from src.codeoceanauxtools.co_fastq import get_prefix
 
 
 # Tests for get_prefix
